backend/requirements.txt
flask
flask_restful
psycopg2==2.9.3
SQLAlchemy
Flask-SQLAlchemy
pyjwt
bcrypt
flask-cors
gunicorn